delete commentary, move all data fn to utility
